backend/python_integration/requirements.txt
web3==6.11.1
gql==3.4.1
python-dotenv==1.0.0
requests==2.31.0